Review remarks implemented
