+++ plugins/tech-radar/src/api.ts
@@ -184,51 +184,3 @@ export interface TechRadarLoaderResponse {
 /*
  * Set up the Radar as a Backstage component.
  */
-
-/**
- * Properties of {@link TechRadarComponent}
- */
-export interface TechRadarComponentProps {
-  /**
-   * ID of this Tech Radar
-   *
-   * @remarks
-   *
-   * Used when there are multiple Tech Radars and passed to {@link TechRadarApi.load}
-   */
-  id?: string;
-  /**
-   * Width of Tech Radar
-   */
-  width: number;
-  /**
-   * Height of Tech Radar
-   */
-  height: number;
-  /**
-   * Custom React props to the `<svg>` element created for Tech Radar
-   */
-  svgProps?: object;
-  /**
-   * Text to filter {@link RadarEntry} inside Tech Radar
-   */
-  searchText?: string;
-}
-
-/**
- * Properties for {@link TechRadarPage}
- */
-export type TechRadarPageProps = TechRadarComponentProps & {
-  /**
-   * Title
-   */
-  title?: string;
-  /**
-   * Subtitle
-   */
-  subtitle?: string;
-  /**
-   * Page Title
-   */
-  pageTitle?: string;
-};
