+++ plugins/tech-radar/src/index.ts
@@ -20,23 +20,22 @@
  * @packageDocumentation
  */
 
+import { RadarPage } from './components';
+
 export {
   techRadarPlugin,
   techRadarPlugin as plugin,
   TechRadarPage,
 } from './plugin';
 
+export * from './components';
+
 /**
  * @deprecated Use plugin extensions instead
  */
-export { RadarPage as Router } from './components/RadarPage';
+export const Router = RadarPage;
 
 /**
  * The TypeScript API for configuring Tech Radar.
  */
 export * from './api';
-
-/**
- * The React component for more advanced use cases.
- */
-export { default as TechRadarComponent } from './components/RadarComponent';
